Add LengthRange filter
Summary: This diff adds a `LengthRange` class, which accepts a string if its length is in a specified range bounded inclusively by two integers. The relevant methods in `Filter.h` and `Filter.cpp`, including test methods, `toString`, and `mergeWith` have been added, and testing code has been added in `FilterTest.cpp` and `ExprToSubfieldFilter.h`. Notably, the `LengthRange` class is only able to merge with `BytesValues` and other `LengthRange` filters, as well as the untyped filters -- no other merges are supported.

Differential Revision: D37836269

fbshipit-source-id: 33fb4fa0a16337484140bd23aacdc59bb36ddc6a