Fix max_by UDF returning empty with non-positive DOUBLE comparison values
The initialValue_ data member of MaxByAggregate is initialized to
std::numeric_limits<U>::min(), which is positive min instead of negative
min in the case of DOUBLE. This results in the max_by UDF returning empty
value if the comparison values/2nd param are all non-positive. This change
is to initialize initialValue_ to std::numeric_limits<U>::lowest()
which is negative min for all types.