Add NegatedBytesRange filter (#2036)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2036

This diff adds a filter for `NOT BETWEEN [lower] AND [upper]` queries on strings (varchar). This is accomplished by creating a wrapper class around a `BytesRange`, which is used to represent the rejected range of values. The corresponding methods for a string filter have been added to `Filter.h` and `Filter.cpp` and the appropriate tests have been added to `FilterTest.cpp`.

Reviewed By: Yuhta

Differential Revision: D37739984

fbshipit-source-id: c9f43ce483adfce5cb639204c7fb2889fb0799aa