Implement MergeWith for NegatedBytesRange filter
Summary: This diff implements the `MergeWith` method for `NegatedBytesRange` filters. Since merging most negated ranges will result in the creation of a `MultiRange` filter that cannot be easily represented otherwise, many of these merges are performed by creating the equivalent `MultiRange` filter with the `toMultiRange` method. This allows the `AND` merge of the new `NegatedBytesRange` filter with another varchar or untyped filter. Tests for this have also been added to `FilterTest.cpp`.

Reviewed By: Yuhta

Differential Revision: D37762978

fbshipit-source-id: 535addcdc32a35ba99dca8e05f719851caf40cc4