Add config to wrap all expressions with try (#1718)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1718

This was an ask from a user.  They want all to catch all exceptions and just return
NULL (the same as what TRY expression provides).  However, wrapping all of their
expressions with TRY(...) made their queries unreadable.  They asked for a setting
to apply the TRY automatically.

This diff adds such a setting to the QueryConfig.  When this setting is applied, the
ExprCompiler wraps all expressions other than TRY itself, as well as field references,
constants and lambdas, with TRY.

Differential Revision: D36686668

fbshipit-source-id: 59c34d2edd7528acdfc0bbd237146411fc1b48bf