Fix bug when inserting into compacted KllSketch (#2037)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2037

After compact (or for a sketch generated by `fromRepeatedValue`), there
is no room in the bottom level, but all levels can still be under-populated.  If
we want to insert into such a sketch, `findLevelToCompact` will not be able to
find a level to shrink.  This fix enables it to first grow the bottom level in this case.

Differential Revision: D37925401

fbshipit-source-id: 559257d747ac8ae9c83b26f190d34ac037b32e86