Add benchmarks for NegatedBytesRange filters
Summary:
This diff adds benchmarks for the `NegatedBytesRange` filters, which are replacing the existing `MultiRange` filters for queries in the form `NOT BETWEEN [lower] AND [upper]` for strings. As such, strings of various lengths are generated and range filters on them are created, which each filter out differing percentages of the values in the test vectors.

The results can be found at P515267564. They indicate that the new NegatedBytesRange filters are an improvement over the old MultiRange filters and run between 1.0x and 1.3x on various inputs, with the speed-up slightly increasing (weak correlation) as the percentage of rejected strings increases and not correlating with the lengths of the strings.

Reviewed By: Yuhta

Differential Revision: D37789402

fbshipit-source-id: 66f2514ecaad2d8bbabde20a7b4a99ef9723a4e3