Skip execution of rows in evalSimplified that have already thrown exceptions (#1455)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1455

Similar to my change here https://github.com/facebookincubator/velox/pull/1081 that did
the same for eval.  (At the time I didn't realize evalSimplified goes through a different path)

Today, when we wrap an expression in a TRY we still evaluate expressions on rows that
have already thrown an exception.

E.g. try(url_extract_path(to_base(4042309829,"c0"))) if to_base throws an exception, we'll
still evaluate url_extract_path on that row

This can cause the system to crash with an SIGSEG for example because the argument
passed in could have a null pointer as it was never initialized.

To address this I've updated evalSimplifiedImpl to unselect any rows that have thrown
exceptions (the TRY will turn the values in these rows to NULL later on anyway).

Differential Revision: D35795650

fbshipit-source-id: c59279538fa3deb28ffd8fb2d0e383a8cedea689