Add Parquet Integer Dictionary
Adds a Parquet RLE  decoder and reading of dictionary page headers.

Adds a special case of non-null that is recognized as a RLE of all ones in single-bit definition levels.

Fixes cases of reading nulls only, e.g. with multiple is null
filters. The generic FormatData::skipNulls gets a flag to indicate if
this is called for a column where only null flags are of interest or
whether this is a part of reading a column with multiple streams.

Does not initialize