Fix JSON cast in TRY expression (#2296)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2296

When evaluating a `TRY(CAST(... as JSON))` expression from MAP or ARRAY
inputs, exceptions from the CAST operation are kept in an error vector in EvalCtx
that will be used to null out the result of the TRY expression. Before this change, the
error vector is automatically populated in EvalCtx::applyToSelectedNoThrow() loops
over the elements of array or map, such that the rows in the error vector doesn't map
to the top-level rows of the array or map vector.

This diff fixes this issue by keeping a reverse mapping from element rows to top-level
rows of the map or array, and manually insert exceptions into the error vector at the
correct position corresponding to the reverse mapping.

Differential Revision: D38546082

fbshipit-source-id: 3fc8f30ca4d1c7b4eec130d12726bdd7c4305755