Fix Final Selection logic in Lambdas.
Summary:
The fix changes how we work with the Final Selection around ExprCallable::apply().
At this border our context changes and the lambda uses different rows, not the rows the context above it was running on.

Imagine 100 rows with a map column and the expression has map_filter(map, lambda).
We were running expression eval on these 100 rows R1 with some rows (maybe) filtered out.

When we descend into the lambda we run its expression (stored in the body_ member).
But we run it on the rows R2, which is received by exploding all elements of maps in all 100 rows - basically we generate R2 based on R1 and how many elements maps have in each row in R1.

Previously we didn't generate finalSelection in the same way and used the original finalSelection from the outer context.
That lead (further down the line) into decoding only R1 number of rows and then processing R2 number of rows with R1 generally being less than R2. Which lead to access uninitialized memory in decoded indices and to invalid results or crashes.

This change converts finalSelection selectivity vector to lambda's space in the same way  we convert the current selectivity vector.

The test also checks that a lazy vector captured in the lambda is being loaded before using it. I believe this needs to be extended by multiple lambdas running on different rows, which we will work right after fixing this issue.

Differential Revision: D39007820

fbshipit-source-id: d250a9a377e6285bee63ad9ea54f6210666c52c5