Implemented format support for week_year token (#2271)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2271

This diff implements support for the week_year token in the DateTimeFormatter.cpp format function.

The following explains week year logic:
The first week of the week year is defined as the first week starting from Monday which contains January 4th of that year. For example, the week year 1970 starts on 1969-12-29, since that is the first week starting from Monday which contains January 4th 1970. In other words:
1969-12-29 -> Monday
1969-12-30 -> Tuesday
1969-12-31 -> Wednesday
1970-01-01 -> Thursday
1970-01-02 -> Friday
1970-01-03 -> Saturday
1970-01-04 -> Sunday

Therefore, all of the dates above belong to the week year 1970, even if the actual year that date resides in is different.

We can also have dates which belong to the previous week year, despite residing in the current year. For example, the week year 2000 starts on January 3rd 2000. This means that 01-01-2000 and 01-02-2000 actually belong to the week year 1999 not 2000.

Differential Revision: D38605915

fbshipit-source-id: b6e18273a677edeca44f5a13ec96eedaa37d8501