Propagate Errors from AsyncSource
Adds error propagation from AsyncSource. If the function of the source
produces an error, the source is considered to have a value and
accessing the value with move() will throw the same error on every
call. move() still returns null if another thread is already waiting
for the source. The waiting thread is still sure to get the error.