Revert "Rename rowGroupMatches to stripeMatches (#2398)"
This reverts commit bf57c9a95facf012e0dc288363ffa3ae8e94cf32.
Stripe is a specific structure in DWRF and ORC, and each stripe has
multiple RowGroups, or called "strides", which are 10000 continuous
rows. Parquet doesn't have Stripe concept, but just RowGroups. A
Parquet file can also be considerred as a file with only 1 Stripe.
Similar to filterRowGroups(), rowGroupMatches() is also meant for
eliminating specific RowGroups in either DWRF or Parquet, NOT Stripes.