Avoid GTEST_SKIP in SpillerTest (#2458)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2458

This change modifies SpillerTest to avoid invoking GTEST_SKIP by clearly defining
the set of Spiller::Types each test supports in the parameter sets.

The advantage of this is that it's at least as clear from looking at the test output
what's being tested and what's not.  In particular, it's clearer in cases where we
weren't using GTEST_SKIP, but rather just returning immediately without testing
anything.

The disadvantage is that it requires explicitly specifying each combination, but at
this point there's only 3 including the set of all Types.

Reviewed By: xiaoxmeng

Differential Revision: D39279876

fbshipit-source-id: e0da5289df5136f87ead8d90ffe16672d6d677c5