Allow mock memory updates
Summary:
We introduced `SimpleMemoryTracker` to accommodate for the trabant memory protection use case where we just rely on a more accurate and simpler accounting for writer coordination.

This diff further splits the behavior of the `SimpleMemoryTracker` and `MemoryUsageTracker` by treating the mock quota shrinks upon shrink reallocation differently, to make `SimpleMemoryTracker` truly accurate.

Reviewed By: nickolay5

Differential Revision: D39422721

fbshipit-source-id: ea0c14325527639493e5fdb36086d16ad4cc4e88