Add large allocation purpose Arena (#2523)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2523

Added MmapArena that does a relatively large mmap ahead of time and manages the mapped memory space for large memory allocations. The goal is to replace existing logic of making a new mmap on every large allocation, which is extremely inefficient and unscalable as kernel can only maintain a certain number of mmaps.

Reviewed By: oerling

Differential Revision: D37131567

fbshipit-source-id: 5fe3a4657a5c72595e95c5a325bcfdae6cc126ab