[CMake] Fix position independent code setting
This patch fixes the name of option, which is used
to set default value for `-fpic` when building velox
code, it should be prefixed with `CMAKE_` to work properly.

Also, since it is a boolean setting, it should be one of
the following:
* 1/0
* TRUE/FALSE
* ON/OFF

Changed to the upper-cased `TRUE` form so that this setting
evaluates to the boolean-true in the cmake execution context.

This change does not affect the CI pipelines, because these
already seem to pass the `-DCMAKE_POSITION_INDEPENDENT_CODE`
manually to the build scripts.

Signed-off-by: Pavel Solodovnikov <pavel.al.solodovnikov@gmail.com>