Save input vector to a file when expression evaluation fails (#2591)
Summary:
When an error occurs during expression evaluation, it is useful to save the
input vector as well as the expression to enable reproducing and debugging the
error in isolation. The expression is already included in the Context: field of
the expression. This change adds logic to save the input vector (EvalCtx::row())
to a file in a directory specified by gflag
velox_save_input_on_expression_failure_path.

This functionality is disabled by default. To enable, specify a path to an existing
writable directory via the velox_save_input_on_expression_failure_path gflag.

```
--velox_save_input_on_expression_failure_path=/tmp
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2591

Reviewed By: laithsakka

Differential Revision: D39645172

Pulled By: mbasmanova

fbshipit-source-id: f7185a8d043571482c68a756f6f1bb700bfd42e9