Addd memory arbitration strategy
Adds a process-wide MemoryManagerStrategy instance and an abstract
base class MemoryConsumer. The strategy keeps a list of consumers.
the consumers each have a top level MemoryUsageTracker that gives
their current usage and limit. The consumers have a method for that
returns an estimate of the memory they can free when requested. The
reclaim() method attempts to free a requested amount of memory. The
actually freed memory is reflected in the tracker.

Adds a TaskMemoryStrategy that arbitrates memory asks between
Tasks. The strategy is invoked via the growCallback of the Task top
level MemoryUsageTracker. The top tracker sees the caps of the Task
level trackers as allocations. It is not a direct parent of these
since its allocated bytes reflects the sum of the caps as opposed to
the sum of the allocations.

TaskMemoryStrategy::reclaim first checks if
the process-wide top tracker has capacity and if so, increases the
limit of the requester by the requested amount. If not enough capacity
is available, this does a dirty read of all running Task's reclaimable
memory and sortrs the potential memory donors largest first. This then
pauses the candidates one after the other and forces enough of them to
release memory to meet the demand. The paused tasks are then resumed.

The requested part of the memory ask is used to increase the limit of
the requester and anything left over is added as available in the top
tracker.  Adds a test with a custom operator that keeps growing a
reservation and can be made to decrese the reservation by contending
users.

In the case of a single Task, when any Driver causes the Task total to
go above Task the cap, the strategy is invoked and since there is only one
Task, the Task is stopped and its operators are made to shrink.

Note that in the test the stop and shrink takes place at the initial
reservation of the MemoryComsumer operator. This is a legitimate place
for calling spill on the requesting operator itself. Once the reservation is obtained, we expect the operator to complete processing its batch before checking for pausing. Whenever an operator goes off-thread as a response to a pause request, it must be in a state where spill() can be safely called on it.

However, if it were to exceed its reservation, then it would invoke
the memory arbitration. This in turn could call spill on the calling
operator, which typically would have bad consequences. For example, a
group by would get spilled in the middle of an update, which could
corrupt its state.  Therefore a spill-protect section must be added as
a follow-up.