Ensure closure of flush policy in writer close (#2651)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2651

In its current state, `flushPolicy_->onClose()` can be skipped when there are exceptions in either flushing the last stripe or closing the writer base.

This diff ensures closure of flush policy when closing the writer. This is an important guarantee for multithreaded coordination.

Future work would be to try and call onClose() in the flush policy destructor instead. Currently if we do that in the multithreaded test suite we see dead locks, and we don't want to invest too much time on making it perfect right now.

Differential Revision: D39842357

fbshipit-source-id: d7d6bf4ba0ef7c6a4de60c33c71114986712be11