Fix velox::functions::jsonExtract() support of wildcard select
Summary:
When there is a wildcard select in the json-path argument, the result of
jsonExtract() should be a folly::dynamic that represents an array.
jsonExtract() right now only generates folly::dynamic of array when the
result contains more than one child, but with wildcard select, arrays with
single child are also possible.

Differential Revision: D39833281

fbshipit-source-id: a504b1c97a43cf38e469886ae3fdbe7f5330f39c