Change the intermediate type of approx_percentile to `ROW` (#18386)
Summary:
X-link: https://github.com/prestodb/presto/pull/18386

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2621

Before this change, intermediate aggregation node for `approx_percentile` in worker only sees a type signature of `VARBINARY -> VARBINARY`, thus not be able to figure out the actual value type and cannot perform the merge properly when the value type is not `DOUBLE`.

This fix changes the intermediate result type to `ROW` so that we can get the value type from it and instantiate accumulator with proper type.

Fix https://github.com/facebookincubator/velox/issues/2430

Reviewed By: mbasmanova

Differential Revision: D39733152

fbshipit-source-id: f7476a9a4a2e92d7fdc976f388a07692787cada8