Fix resizing bug in DecodedVector wrapping a ConstantVector
Summary:
The wrapping code in DecodedVector currently short circuits when the input vector that is being wrapped is constant encoded. This condition in itself is insufficient because the size of the input vector is being ignored and returned as is, irrespective of the size of the selection.

In this change, I updated the short circuit condition to reflect the size check as the resizing of constant vector logic already exists in dictionary wrapping logic. Added a test to show this condition and validated that it fails on size check before the fix.

Differential Revision: D39917906

fbshipit-source-id: 2cb0bf0379707b5ea8f7e2a320f33c341b229eaa