Make peer values offset from the beginning of the partition
Peer value offsets in the WindowFunction::apply(...) calls were
row number values from the sortedRows_ version of the RowContainer (where
all the input rows are stored). In the WindowFunction::apply(...) values
only make sense from the start of the partition that is being processed.

So changing the peer values calculation with that logic.